Hopefully fixed crashing issue
The crashing issue is an unexpected error with no helpful description
that happens seemingly randomly. I was able to confirm that the issue
was introduced in either commit 2cabd33 or earlier. I haven’t been able
to reproduce the bug in commit 4230981, so I’m assuming it was a
divide-by-zero error in the speed2() function of the Body class.