QuadTree improvements
Point class also is able to convert to string in more situations
implicitly, such as when printing a list of points. Also added some
benchmarking code to test the QuadTree's speed. Added a dictionary
indexed using objects and which returns the nodes colliding with that
object (in order to see if the object's position should be re-evaluated
in the tree)